+++ CHANGELOG.md
@@ -12,12 +12,11 @@ All notable changes to this project will be documented in this file.
 -->
 
 ## [Unreleased]
+## [7.2.1]
 
 ### Added
 
 -   Added support for vector-based layers for basemaps, currently only available for the custom basemap.
-
-## [7.2.1]
 
 ### Changed